+++ src/pystackapi/__init__.py
@@ -1,5 +1,6 @@
+from .chat import Chat
 from .network import Network
 from .site import Site
 
 __version__ = '0.3.0dev'
-__all__ = ('Site', 'Network', 'sites', 'errors', 'item')
+__all__ = ('Chat', 'Site', 'Network', 'sites', 'errors', 'item')